backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
python-multipart

fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
passlib
python-jose
google-generativeai
requests
email-validator